Title: ya_get_dataset()

Description: Gathers information about the datasets in BPstat.

Usage

ya_get_dataset(lang = "EN")

Arguments

lang

A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"

Value

A data table with the information about the BPstat domains

Examples

ya_get_dataset()