Title: ya_get_dimension_item()

Description: A function that gathers information available in BPstat about the dataset items.

Usage

ya_get_dimension_item(lang = "EN")

Arguments

lang

A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"

Value

A data table with the information in a more human readable form with a search option using regex.

Examples

ya_get_dimension_item()