Title: ya_info_serie()

Description: A function allows to user to plot, see in a table or reuse in other context, data from a determined BPstat serie ID.

Usage

ya_info_serie(series_id = 0, operation = "P", lang = "EN")

Arguments

series_id

A serie ID from BPstat. Defaults to 0 (zero)

operation

A string with the length of 1 (one), representing the following operations: P - plot T - table Df - dataframe

lang

A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"

Value

From the user choice can be a plot, a table or a tibble for further analysis

Examples

ya_info_serie(series_id = 12528985)