Title: ya_my_favourites()

Description: this function allows user to add, delete and view favourite series from BPStat

Usage

ya_my_favourites(series_id = 0, operation = "V", lang = "EN")

Arguments

series_id

A serie ID from BPstat. Defaults to 0 (zero)

operation

A string with the length of 1 (one), representing the following operations: A - add D - delete V - view

lang

A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"

Value

A file named «ya_favourites» with user favourite series

Examples

ya_my_favourites()