Title: ya_search()

Description: This is a function that allows to the user search directly de BPstat data through the API.

Usage

ya_search(query, lang = "EN")

Arguments

query

A string that represents a query to be executed in BPstat

lang

A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"

Value

A data table with the query results from BPstat.

Examples

ya_search(query="ihpc")